requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-swiftclient==4.4.0
PyMySQL==1.1.0
django-environ==0.11.2
Pillow==10.1.0
python-magic==0.4.27
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0